+++ changelog.md
@@ -78,3 +78,4 @@
 * Change from `isRoot`, `isPublic` etc columns in user and role models to single column `type`
 * Create 'admin' and 'user' roles
 * Refactor db stocking code
+* Controllers for permissions